# Add performance profiling for slow `GetDutiesV2` operations