--- a/changelog/ttsao_add-duties-profiling.md
+++ b/changelog/ttsao_add-duties-profiling.md
@@ -1,0 +1,3 @@
+### Added
+
+- Performance profiling for GetDutiesV2 operations taking over 2 seconds