This PR adds automatic performance profiling when the GetDutiesV2 RPC endpoint takes longer than 2 seconds to respon

### Changes
  - New flag: `--slow-duties-profile` - Boolean flag to enable automatic profiling
  - Profiling trigger: Automatically starts profiling when GetDutiesV2 takes >2s
  - Profile types captured:
    - CPU profile
    - Mutex contention profile
    - Goroutine profile
    - Heap memory profile
  - Storage location: Profiles saved to <datadir>/debug/ directory

### Examples
```
Aug 03 07:25:26 bear beacon-chain[817046]: time="2025-08-03 07:25:26.64" level=warning msg="GetDutiesV2 taking longer than 2s, capturing profiles" currentDuration=2.00051592s epoch=383873 numValidators=30000 prefix="rpc/validator" profileDir="/home/t/consensus-db/debug"
Aug 03 07:25:26 bear beacon-chain[817046]: time="2025-08-03 07:25:26.65" level=info msg="Mutex profile captured" file="/home/t/consensus-db/debug/mutex-duties-20250803-072526.prof" prefix="rpc/validator"
Aug 03 07:25:26 bear beacon-chain[817046]: time="2025-08-03 07:25:26.65" level=info msg="Goroutine profile captured" file="/home/t/consensus-db/debug/goroutine-duties-20250803-072526.prof" prefix="rpc/validator"
Aug 03 07:25:26 bear beacon-chain[817046]: time="2025-08-03 07:25:26.82" level=info msg="Heap profile captured" file="/home/t/consensus-db/debug/heap-duties-20250803-072526.prof" prefix="rpc/validator"

<Profiles captured later>

```

```
Batch Size   Duration   Status     Response Size   Notes
--------------------------------------------------------------------------------
10,000       4.337      SUCCESS    6,483,477       
20,000       6.414      SUCCESS    12,984,180      
30,000       9.641      SUCCESS    19,489,487      


ls
cpu-duties-20250803-072514.prof        heap-duties-20250803-072514.prof
cpu-duties-20250803-072519.prof        heap-duties-20250803-072519.prof
cpu-duties-20250803-072526.prof        heap-duties-20250803-072526.prof
goroutine-duties-20250803-072514.prof  mutex-duties-20250803-072514.prof
goroutine-duties-20250803-072519.prof  mutex-duties-20250803-072519.prof
goroutine-duties-20250803-072526.prof  mutex-duties-20250803-072526.prof
```


<img width="1692" height="399" alt="Screenshot 2025-08-03 at 7 48 28 AM" src="https://github.com/user-attachments/assets/e3da33cb-d979-4a83-b406-72d20d53b84b" />


#### What we have found so far
**Profile bottleneck**
  - AssignmentForValidator consumes 41-68% of CPU time
  - This function's CPU usage increased significantly between captures (2.12s → 4.02s → 6.94s)

**Heap bottleneck**
  - BeaconStateElectra.UnmarshalSSZ: 585MB (19% of heap)
  - ValidatorIndexMap: 531MB (17% of heap)
  - Validator.UnmarshalSSZ: 329MB (11% of heap)

**Mutex Contention**
  - Mutex unlock operations causing 24.49s of blocking time mostly on p2p
    - github.com/OffchainLabs/prysm/v6/beacon-chain/p2p.(*Service).pingPeersAndLogEnr - 14.32s
    - github.com/OffchainLabs/prysm/v6/beacon-chain/sync.(*Service).sendPingRequest - 16.18s
    - github.com/libp2p/go-libp2p/p2p/host/basic.(*streamWrapper).Read - 16.55s
    - github.com/multiformats/go-multistream.(*lazyClientConn).Read - 16.55s
    - github.com/OffchainLabs/prysm/v6/beacon-chain/sync.(*Service).reValidatePeer - 2.02s
    - github.com/OffchainLabs/prysm/v6/beacon-chain/sync.(*Service).sendMetaDataRequest - 1.66s